+++ config/routes.rb
@@ -1,7 +1,11 @@
 Rails.application.routes.draw do
-  resources :categories, :only => [:index]
-  resources :payments, :except => [:new, :edit], :defaults => {:format => 'json'}
-  get '/settlement' => 'payments#settle', :defaults => {:format => 'json'}
-  get '/login' => 'login#form'
+  namespace :api, :format => 'json' do
+    resources :categories, :only => [:index]
+    resources :payments, :except => [:new, :edit]
+    get '/settlement' => 'payments#settle'
+  end
+
+  resources :payments, :only => [:index], :format => 'html'
+  get '/login' => 'login#form', :format => 'html'
   post '/login' => 'login#authenticate_user'
 end