+++ spec/controllers/api/category/index_spec.rb
@@ -1,11 +1,11 @@
 # coding: utf-8
 require 'rails_helper'
 
-describe CategoriesController, :type => :controller do
+describe Api::CategoriesController, :type => :controller do
   shared_context 'カテゴリを検索する' do |param = {}, app_auth_header = CommonHelper.app_auth_header|
     before(:all) do
       client.header('Authorization', app_auth_header)
-      @res = client.get('/categories.json', param)
+      @res = client.get('/api/categories', param)
       @pbody = JSON.parse(@res.body) rescue nil
     end
   end