+++ spec/support/requests/shared/payment.rb
@@ -1,16 +1,16 @@
 # coding: utf-8
-shared_context 'POST /payments' do |payment|
+shared_context 'POST /api/payments' do |payment|
   before(:all) do
     header = {'Authorization' => app_auth_header}.merge(content_type_json)
-    @res = http_client.post("#{base_url}/payments", {:payments => payment}.to_json, header)
+    @res = http_client.post("#{base_url}/api/payments", {:payments => payment}.to_json, header)
     @pbody = JSON.parse(@res.body) rescue nil
   end
 end
 
-shared_context 'GET /payments' do |params = {}|
+shared_context 'GET /api/payments' do |params = {}|
   before(:all) do
     header = {'Authorization' => app_auth_header}
-    @res = http_client.get("#{base_url}/payments", params.to_json, header)
+    @res = http_client.get("#{base_url}/api/payments", params.to_json, header)
     @pbody = JSON.parse(@res.body) rescue nil
   end
 end